Extract BeaconSummaryEditing state component + tests
